Refactor InventoryFormat into Inventory again. No need to have virtual implementation
