+++ objects.mk
@@ -2,4 +2,5 @@ USER_OBJS :=
 
 LIBS := -lz -lcrypto -lssl
 
+CFLAGS := -fPIE
 CXXFLAGS := $(CFLAGS) -O2 -Wall -I ./
